requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-annoying==0.10.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-localflavor==2.0
django-material==1.2.5
django-viewflow==1.2.5
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
